refactor: rename ProjectApi to ProjectsApi
